pyproject.toml
[tool.poetry]
name = "llm-calc"
version = "0.1.0"
description = ""
authors = ["Alex Goodell <alexgoodell@gmail.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
jupyter = "^1.0.0"
e2b-code-interpreter = "^0.0.10"
langchain = "^0.3.3"
langchainhub = "^0.1.20"
langchain-openai = "^0.2.2"
pandas = "^2.2.2"
load-dotenv = "^0.1.0"
pandantic = "^0.3.0"
duckdb = "^1.0.0"
ibis-framework = {extras = ["duckdb"], version = "^9.1.0"}
tabulate = "^0.9.0"
rich = "^13.7.1"
lxml = "^5.2.2"
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
openpyxl = "^3.1.5"
jupysql = "^0.10.12"
matplotlib = "^3.9.1"
duckdb-engine = "^0.13.0"
toml = "^0.10.2"
langchain-huggingface = "^0.1.0"
text-generation = "^0.7.0"
transformers = "^4.42.4"
google-search-results = "^2.4.2"
numexpr = "^2.10.1"
sentencepiece = "^0.2.0"
jinja2 = "^3.1.4"
langchain-groq = "^0.2.0"
langchain-community = "^0.3.2"
beautifulsoup4 = "^4.12.3"
pypandoc-binary = "^1.13"
langchain-chroma = "^0.1.2"
unstructured = "^0.15.0"
python-magic = "^0.4.27"
langgraph = "^0.2.35"
pytest = "^8.3.1"
inquirerpy = "^0.3.4"
ipython = "^8.26.0"
typer = "^0.12.3"
wonderwords = "^2.2.0"
langsmith = "^0.1.98"
markdown = "^3.6"
invoke = "^2.2.0"
black = "^24.8.0"
dirtyjson = "^1.0.8"
ofunctions = "^2.7.0"
langchain-anthropic = "^0.2.3"
langchain-together = "^0.2.0"
langchain-experimental = "^0.3.2"
llamaapi = "^0.1.36"
fastapi = {extras = ["standard"], version = "^0.114.0"}
llm = "^0.16"
llm-claude-3 = "^0.4.1"
outlines = "^0.0.46"
langchain-core = "^0.3.10"
langchain-fireworks = "^0.2.1"
faker = "^30.8.1"
plotly = "^5.24.1"
matplotlib-inline = "^0.1.7"
kaleido = "0.2.1"
xarray = "^2024.10.0"
seaborn = "^0.13.2"
statsmodels = "^0.14.4"
bioinfokit = "^2.1.4"
chart-studio = "^1.1.0"
onnxruntime = "1.18.1"
torch = "2.0.1"
gspread = "^6.1.4"
gspread-dataframe = "^4.0.0"
jedi = "^0.19.2"

[tool.poetry.scripts]
llmcalc = "llm_calc.app:app"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"
